requirements.txt
Django==1.11.7
django-debug-toolbar==1.9.1
django-multiselectfield==0.1.8
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytz==2017.3
sqlparse==0.2.4
whitenoise==3.3.1
